src/main/webapp/app/entities/atividade-empresa/delete/atividade-empresa-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { AtividadeEmpresaService } from '../service/atividade-empresa.service';

import { AtividadeEmpresaDeleteDialogComponent } from './atividade-empresa-delete-dialog.component';

describe('AtividadeEmpresa Management Delete Component', () => {
  let comp: AtividadeEmpresaDeleteDialogComponent;
  let fixture: ComponentFixture<AtividadeEmpresaDeleteDialogComponent>;
  let service: AtividadeEmpresaService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [AtividadeEmpresaDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(AtividadeEmpresaDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(AtividadeEmpresaDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(AtividadeEmpresaService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
